lesson-18-ai/GenAI/phate-tutorial/requirements.txt
phate 
umap-learn 
pacmap 
trimap
matplotlib 
scikit-learn